requirements.txt
autopep8
python-dotenv

pandas
numpy
matplotlib

yfinance
selenium
webdriver_manager